+++ Godeps/_workspace/src/github.com/macaron-contrib/session/nodb/nodb.go
@@ -1,5 +1,4 @@
-// Copyright 2013 Beego Authors
-// Copyright 2014 Unknwon
+// Copyright 2015 Unknwon
 //
 // Licensed under the Apache License, Version 2.0 (the "License"): you may
 // not use this file except in compliance with the License. You may obtain
@@ -17,29 +16,26 @@ package session
 
 import (
 	"fmt"
-	"strings"
 	"sync"
 
-	"github.com/Unknwon/com"
-	"github.com/siddontang/ledisdb/config"
-	"github.com/siddontang/ledisdb/ledis"
-	"gopkg.in/ini.v1"
+	"github.com/lunny/nodb"
+	"github.com/lunny/nodb/config"
 
 	"github.com/macaron-contrib/session"
 )
 
-// LedisStore represents a ledis session store implementation.
-type LedisStore struct {
-	c      *ledis.DB
+// NodbStore represents a nodb session store implementation.
+type NodbStore struct {
+	c      *nodb.DB
 	sid    string
 	expire int64
 	lock   sync.RWMutex
 	data   map[interface{}]interface{}
 }
 
-// NewLedisStore creates and returns a ledis session store.
-func NewLedisStore(c *ledis.DB, sid string, expire int64, kv map[interface{}]interface{}) *LedisStore {
-	return &LedisStore{
+// NewNodbStore creates and returns a ledis session store.
+func NewNodbStore(c *nodb.DB, sid string, expire int64, kv map[interface{}]interface{}) *NodbStore {
+	return &NodbStore{
 		c:      c,
 		expire: expire,
 		sid:    sid,
@@ -48,7 +44,7 @@ func NewLedisStore(c *ledis.DB, sid string, expire int64, kv map[interface{}]int
 }
 
 // Set sets value to given key in session.
-func (s *LedisStore) Set(key, val interface{}) error {
+func (s *NodbStore) Set(key, val interface{}) error {
 	s.lock.Lock()
 	defer s.lock.Unlock()
 
@@ -57,7 +53,7 @@ func (s *LedisStore) Set(key, val interface{}) error {
 }
 
 // Get gets value by given key in session.
-func (s *LedisStore) Get(key interface{}) interface{} {
+func (s *NodbStore) Get(key interface{}) interface{} {
 	s.lock.RLock()
 	defer s.lock.RUnlock()
 
@@ -65,7 +61,7 @@ func (s *LedisStore) Get(key interface{}) interface{} {
 }
 
 // Delete delete a key from session.
-func (s *LedisStore) Delete(key interface{}) error {
+func (s *NodbStore) Delete(key interface{}) error {
 	s.lock.Lock()
 	defer s.lock.Unlock()
 
@@ -74,12 +70,12 @@ func (s *LedisStore) Delete(key interface{}) error {
 }
 
 // ID returns current session ID.
-func (s *LedisStore) ID() string {
+func (s *NodbStore) ID() string {
 	return s.sid
 }
 
 // Release releases resource and save data to provider.
-func (s *LedisStore) Release() error {
+func (s *NodbStore) Release() error {
 	data, err := session.EncodeGob(s.data)
 	if err != nil {
 		return err
@@ -93,7 +89,7 @@ func (s *LedisStore) Release() error {
 }
 
 // Flush deletes all session data.
-func (s *LedisStore) Flush() error {
+func (s *NodbStore) Flush() error {
 	s.lock.Lock()
 	defer s.lock.Unlock()
 
@@ -101,45 +97,29 @@ func (s *LedisStore) Flush() error {
 	return nil
 }
 
-// LedisProvider represents a ledis session provider implementation.
-type LedisProvider struct {
-	c      *ledis.DB
+// NodbProvider represents a ledis session provider implementation.
+type NodbProvider struct {
+	c      *nodb.DB
 	expire int64
 }
 
-// Init initializes ledis session provider.
-// configs: data_dir=./app.db,db=0
-func (p *LedisProvider) Init(expire int64, configs string) error {
+// Init initializes nodb session provider.
+func (p *NodbProvider) Init(expire int64, configs string) error {
 	p.expire = expire
 
-	cfg, err := ini.Load([]byte(strings.Replace(configs, ",", "\n", -1)))
+	cfg := new(config.Config)
+	cfg.DataDir = configs
+	dbs, err := nodb.Open(cfg)
 	if err != nil {
-		return err
+		return fmt.Errorf("session/nodb: error opening db: %v", err)
 	}
 
-	db := 0
-	opt := new(config.Config)
-	for k, v := range cfg.Section("").KeysHash() {
-		switch k {
-		case "data_dir":
-			opt.DataDir = v
-		case "db":
-			db = com.StrTo(v).MustInt()
-		default:
-			return fmt.Errorf("session/ledis: unsupported option '%s'", k)
-		}
-	}
-
-	l, err := ledis.Open(opt)
-	if err != nil {
-		return fmt.Errorf("session/ledis: error opening db: %v", err)
-	}
-	p.c, err = l.Select(db)
+	p.c, err = dbs.Select(0)
 	return err
 }
 
 // Read returns raw session store by session ID.
-func (p *LedisProvider) Read(sid string) (session.RawStore, error) {
+func (p *NodbProvider) Read(sid string) (session.RawStore, error) {
 	if !p.Exist(sid) {
 		if err := p.c.Set([]byte(sid), []byte("")); err != nil {
 			return nil, err
@@ -160,23 +140,23 @@ func (p *LedisProvider) Read(sid string) (session.RawStore, error) {
 		}
 	}
 
-	return NewLedisStore(p.c, sid, p.expire, kv), nil
+	return NewNodbStore(p.c, sid, p.expire, kv), nil
 }
 
 // Exist returns true if session with given ID exists.
-func (p *LedisProvider) Exist(sid string) bool {
+func (p *NodbProvider) Exist(sid string) bool {
 	count, err := p.c.Exists([]byte(sid))
 	return err == nil && count > 0
 }
 
 // Destory deletes a session by session ID.
-func (p *LedisProvider) Destory(sid string) error {
+func (p *NodbProvider) Destory(sid string) error {
 	_, err := p.c.Del([]byte(sid))
 	return err
 }
 
 // Regenerate regenerates a session store from old session ID to new one.
-func (p *LedisProvider) Regenerate(oldsid, sid string) (_ session.RawStore, err error) {
+func (p *NodbProvider) Regenerate(oldsid, sid string) (_ session.RawStore, err error) {
 	if p.Exist(sid) {
 		return nil, fmt.Errorf("new sid '%s' already exists", sid)
 	}
@@ -189,7 +169,10 @@ func (p *LedisProvider) Regenerate(oldsid, sid string) (_ session.RawStore, err 
 			return nil, err
 		}
 	}
-	if err = p.c.SetEX([]byte(sid), p.expire, kvs); err != nil {
+
+	if err = p.c.Set([]byte(sid), kvs); err != nil {
+		return nil, err
+	} else if _, err = p.c.Expire([]byte(sid), p.expire); err != nil {
 		return nil, err
 	}
 
@@ -203,20 +186,18 @@ func (p *LedisProvider) Regenerate(oldsid, sid string) (_ session.RawStore, err 
 		}
 	}
 
-	return NewLedisStore(p.c, sid, p.expire, kv), nil
+	return NewNodbStore(p.c, sid, p.expire, kv), nil
 }
 
 // Count counts and returns number of sessions.
-func (p *LedisProvider) Count() int {
+func (p *NodbProvider) Count() int {
 	// FIXME: how come this library does not have DbSize() method?
 	return -1
 }
 
 // GC calls GC to clean expired sessions.
-func (p *LedisProvider) GC() {
-	// FIXME: wtf???
-}
+func (p *NodbProvider) GC() {}
 
 func init() {
-	session.Register("ledis", &LedisProvider{})
+	session.Register("nodb", &NodbProvider{})
 }