+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 2.0.0-RC1  (unreleased)
 
+**FIxes**
+- [Issue #1681](https://github.com/grafana/grafana/issues/1681). MySQL session: fixed problem using mysql as session store
 - [Issue #1671](https://github.com/grafana/grafana/issues/1671). Data sources: Fixed issue with changing default data source (should not require full page load to take effect, now fixed)
 
 # 2.0.0-Beta1 (2015-03-30)